requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==5.1.0
pytz==2024.1
plotly==5.20.0 